Add variable flexibility of `ListEntry`
Fully fixes any `ListEntry` jank once and for all (hopefully).
Before, all list entries had the left side
widget wrapped in a `Flexible` widget.
This caused a problem with SettingsListEntry, because here, the
"variable"-sized widget (e.g. user email text) was on the right side.
The previous solution for that problem was to place the right side
widget in a fixed size constrained box, but that fix still doesn't play
well with variable length emails.
Now, `sideToExpand` tells the ListEntry which side should be displayed
in full, and which side should be wrapped in a `Flexible` widget.

Misc: Used correct `SettingDescription` widget for the right side of
each `SettingListEntry`.