refactor(ui): add RoundedButton.bright constructor (#555)
This commit reduces code duplication by creating a new constructor to
the
`RoundedButton` class, making it usable as a replacement for the
following buttons:

1. the "Resend verification email" button
2. the button in `SplashErrorPage`

Beforehand, these two buttons used tedious boilerplate by invoking the
Material `ElevatedButton`; a thing the `RoundedButton` was supposed to
eliminate.

Additionally, I simplified some code in the files this change affected.